+++ DatadogSessionReplay/Sources/Recorder/ViewTreeSnapshotProducer/ViewTreeSnapshot/NodeRecorders/UILabelRecorder.swift
@@ -36,24 +36,15 @@ internal class UILabelRecorder: NodeRecorder {
             return InvisibleElement.constant
         }
 
-        // The actual frame of the text, which is smaller than the frame of the label:
-        let textFrame = CGRect(
-            x: attributes.frame.minX,
-            y: attributes.frame.minY + (attributes.frame.height - attributes.intrinsicContentSize.height) * 0.5,
-            width: attributes.intrinsicContentSize.width,
-            height: attributes.intrinsicContentSize.height
-        )
-
         let builder = UILabelWireframesBuilder(
             wireframeID: context.ids.nodeID(for: label),
             attributes: attributes,
             text: label.text ?? "",
             textColor: label.textColor?.cgColor,
-            textAlignment: nil,
+            textAlignment: label.textAlignment,
             font: label.font,
             fontScalingEnabled: label.adjustsFontSizeToFitWidth,
-            textObfuscator: textObfuscator(context),
-            wireframeRect: textFrame
+            textObfuscator: textObfuscator(context)
         )
         let node = Node(viewAttributes: attributes, wireframesBuilder: builderOverride(builder))
         return SpecificElement(subtreeStrategy: .ignore, nodes: [node])
@@ -69,7 +60,7 @@ internal struct UILabelWireframesBuilder: NodeWireframesBuilder {
     /// The color of the text.
     var textColor: CGColor?
     /// The alignment of the text.
-    var textAlignment: SRTextPosition.Alignment?
+    var textAlignment: NSTextAlignment
     /// The font used by the label.
     let font: UIFont?
     /// Flag that determines if font should be scaled
@@ -77,16 +68,17 @@ internal struct UILabelWireframesBuilder: NodeWireframesBuilder {
     /// Text obfuscator for masking text.
     let textObfuscator: TextObfuscating
 
-    let wireframeRect: CGRect
+    var wireframeRect: CGRect {
+        attributes.frame
+    }
 
     func buildWireframes(with builder: WireframesBuilder) -> [SRWireframe] {
         return [
             builder.createTextWireframe(
                 id: wireframeID,
-                frame: attributes.frame,
+                frame: wireframeRect,
                 text: textObfuscator.mask(text: text),
-                textFrame: wireframeRect,
-                textAlignment: textAlignment,
+                textAlignment: .init(systemTextAlignment: textAlignment),
                 textColor: textColor,
                 font: font,
                 fontScalingEnabled: fontScalingEnabled,