+++ Sources/Datadog/Tracer.swift
@@ -65,6 +65,8 @@ public class Tracer: OTTracer {
 
     internal let activeSpansPool = ActiveSpansPool()
 
+    internal let sampler: Sampler
+
     /// Tracer Attributes shared with other Feature registered in core.
     internal struct Attributes {
         internal static let traceID = "dd.trace_id"
@@ -146,6 +148,7 @@ public class Tracer: OTTracer {
         self.dateProvider = dateProvider
         self.rumIntegration = rumIntegration
         self.loggingIntegration = loggingIntegration
+        self.sampler = Sampler(samplingRate: configuration.samplingRate)
     }
 
     // MARK: - Open Tracing interface