+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+# 1.19.0 / 26-04-2023
+- [BUGFIX] Fix view attributes override by action attributes. See [#1250][]
+- [IMPROVEMENT] Add Tracer sampling rate. See [#1259][]
+- [BUGFIX] Fix RUM context not being attached to log when no user action exists. See [#1264][]
+
 # 1.18.0 / 19-04-2023
 - [IMPROVEMENT] Add start reason to the session. See [#1247][]
 - [IMPROVEMENT] Add ability to stop the session. See [#1219][]
@@ -452,6 +457,9 @@
 [#1219]: https://github.com/DataDog/dd-sdk-ios/pull/1219
 [#1220]: https://github.com/DataDog/dd-sdk-ios/pull/1220
 [#1247]: https://github.com/DataDog/dd-sdk-ios/pull/1247
+[#1250]: https://github.com/DataDog/dd-sdk-ios/pull/1250
+[#1259]: https://github.com/DataDog/dd-sdk-ios/pull/1259
+[#1264]: https://github.com/DataDog/dd-sdk-ios/pull/1264
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu